MISC/ANAS/openCV/try00-read/requirements.txt
blinker==1.6.2
caer==2.0.8
click==8.1.6
contourpy==1.1.0
cycler==0.11.0
Flask==2.3.2
Flask-WTF==1.1.1
fonttools==4.41.1
itsdangerous==2.1.2
Jinja2==3.1.2
kiwisolver==1.4.4
MarkupSafe==2.1.3
matplotlib==3.7.2
mypy==1.4.1
mypy-extensions==1.0.0
numpy==1.25.2
opencv-contrib-python==4.8.0.74
packaging==23.1
Pillow==10.0.0
pyparsing==3.0.9
python-dateutil==2.8.2
six==1.16.0
tomli==2.0.1
typing_extensions==4.7.1
Werkzeug==2.3.6
WTForms==3.0.1
